Move clustering queries to own file
